Organize fetches and update path
