+++ src/Components/IndividualTrail/IndividualTrail.js
@@ -4,7 +4,7 @@ import './IndividualTrail.css'
 import { useSelector, useDispatch } from 'react-redux'
 import { NavLink } from 'react-router-dom'
 import { setTrail, setUser } from "../../actions";
-import { postToFavorites } from "../../utilis/apiCalls"
+import { postToFavorites } from "../../utilities/apiCalls"
 
 function IndividualTrail() {
   const selectedTrail = useSelector((state) => state.selectedTrail.data.attributes)
