--- src/apiCalls/getAllTrails.js
@@ -1,8 +1,0 @@
-export const getAllTrails = async () => {
-  const url = "http://localhost:3000/api/v1/trails";
-  const response = await fetch(url);
-  if (response.status < 200 && response.status > 299) {
-    throw response;
-  }
-  return response.json();
-};
